Re-organize App.Http, split pure code out and tidy up.
  The module was starting to get big with many level of abstractions
  intertwined. For the sake of keeping the separation good (and
  possibly, indidually improve test coverage of the sub-parts), I've
  move all the pure helpers into separate modules under the 'Data'
  section. Remains in 'App.Http' only the actual route handlers. This is
  a bit cleaner and allow to extend even more the http endpoints while
  keeping things clean and tidy.
